requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0

# PostgreSQL драйверы
asyncpg==0.29.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
alembic==1.12.1

# Для миграций
alembic==1.12.1

# Дополнительно
httpx==0.25.1
python-multipart==0.0.6